ACCUMULO-3455: ShellOptionsJC should extend ClientOpts
The ticket seems outdated and while there is some overlap between the two classes, their use looks different enough to keep them separated. So I just unified their password handling.
